+++ src/pages/main.tsx
@@ -1,6 +1,6 @@
-import HistoryCard from "@/components/card/HistoryCard";
-import ProfileCard from "@/components/card/ProfileCard";
-import StatusCard from "@/components/card/StatusCard";
+import HistoryCard from "@/card/main/HistoryCard";
+import ProfileCard from "@/card/main/ProfileCard";
+import StatusCard from "@/card/main/StatusCard";
 import Container from "@/layouts/Container";
 import FlexBox from "@/layouts/FlexBox";
 import TopNav from "@/layouts/TopNav";
